Stub in an instructlab-taxonomy datasetio provider
This adds an instructlab-taxonomy datasetio provider. The job of this
provider is to represent a taxonomy git repository as a dataset, which
is kind of hacked together right now to just generate a dataset with
`qna_path` and `qna_contents` fields where `qna_path` is the relative
path to that qna.yaml from the taxonomy root and `qna_contents` is the
literal string contents of that qna.yaml file.

Signed-off-by: Ben Browning <bbrownin@redhat.com>